Lots of enhancements.  The major changes include preliminary interop work (server, client & fixes), the ability to disable all use of wchar_t, and miscellaneous warning cleanups.
git-svn-id: http://svn.code.sf.net/p/xmlrpc-c/code@87 98333e67-4a24-44d7-a75c-e53540dd3050

+++ trunk/xmlrpc-c/examples/meerkat-app-list.cc
@@ -41,7 +41,7 @@ static void list_apps (int hours) {
     // Print our results.
     int first = 1;
     size_t app_count = apps.arraySize();
-    for (int i = 0; i < app_count; i++) {
+    for (size_t i = 0; i < app_count; i++) {
 	XmlRpcValue app = apps.arrayGetItem(i);
 
 	// Get some information about our application.
